mensagem de erro adicionada

+++ usuario/cadastros_manager.php
@@ -13,6 +13,7 @@ require_once __DIR__ . '/../db/entities/pagar.php';
 require_once __DIR__ . '/../db/entities/centrocustos.php';
 require_once __DIR__ . '/../db/entities/banco01.php';
 require_once __DIR__ . '/../db/entities/banco02.php';
+require_once __DIR__ . '/../db/entities/fecha01.php';
 
 session_start();
 
@@ -457,7 +458,12 @@ if (isset($view) && $view == 'cadastro') {
             
         } else if (isset($acao) && $acao == 'excluir') {
             
+            
             if(!Con02::read(null, null, $id_conta) ) {
+                if(Fecha01::read(id_empresa: $_SESSION['usuario']->id_empresa, id_titulo: $id_conta)) {
+                    header('Location: contas.php?con01id='.$id_conta.'&erro=usado');
+                    exit;
+                }
                 Con01::delete($id_conta);
             } else {
                 header('Location: contas.php?con01id='.$id_conta.'&erro=usado_sub');
@@ -492,6 +498,10 @@ if (isset($view) && $view == 'cadastro') {
             } else if (isset($acao) && $acao == 'excluir') {
                 $con02 = Con02::read($id_conta02, idempresa:$_SESSION['usuario']->id_empresa)[0];
                 $con01 = Con01::read($con02->id_con01, idempresa:$_SESSION['usuario']->id_empresa)[0];
+                if(Fecha01::read(id_empresa: $_SESSION['usuario']->id_empresa, id_subtitulo: $con02->id)) {
+                    header('Location: contas.php?con01id='.$id_conta.'&erro=usado');
+                    exit;
+                }
                 if($con01->tipo == 'C') {
 
                     if(!Rec02::read(id_empresa:$_SESSION['usuario']->id_empresa, filtro_con02: $id_conta02) && !Ban02::read(id_empresa: $_SESSION['usuario']->id_empresa, filtro_subtitulo: $id_conta02)) {
